Implement revocations.
