+++ lib/http.js
@@ -3,36 +3,22 @@
  */
 'use strict';
 
-const asyncHandler = require('express-async-handler');
-const base58 = require('bs58');
+const {asyncHandler} = require('bedrock-express');
 const bedrock = require('bedrock');
 const brPassport = require('bedrock-passport');
 const brZCapStorage = require('bedrock-zcap-storage');
-const {config} = bedrock;
+const {config, util: {BedrockError}} = bedrock;
 const cors = require('cors');
-const crypto = require('crypto');
 const database = require('bedrock-mongodb');
+const helpers = require('./helpers');
 const jsigs = require('jsonld-signatures');
-const {verifyCapabilityInvocation} = require('http-signature-zcap-verify');
 const storage = require('./storage');
 const {validate} = require('bedrock-validation');
-const {promisify} = require('util');
-const getRandomBytes = promisify(crypto.randomBytes);
-const {extendContextLoader, SECURITY_CONTEXT_V2_URL} = jsigs;
+const {extendContextLoader} = jsigs;
 const {Ed25519Signature2018} = jsigs.suites;
 const {CapabilityDelegation} = require('ocapld');
 require('bedrock-express');
-const {
-  ensureAuthenticated
-} = brPassport;
-const {BedrockError} = bedrock.util;
-
-// load config defaults
-require('./config');
-
-// module API
-const api = {};
-module.exports = api;
+const {ensureAuthenticated} = brPassport;
 
 // Note: edv routes are not configurable per the spec
 const routes = {
@@ -51,7 +37,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
   const baseStorageUrl = `${baseUri}${routes.edvs}`;
 
   function _getEdvId(edvIdParam) {
-    _assert128BitId(edvIdParam);
+    helpers.assert128BitId(edvIdParam);
     return `${baseStorageUrl}/${edvIdParam}`;
   }
 
@@ -75,7 +61,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       const {actor} = (req.user || {});
       delete req.body.id;
-      const id = _getEdvId(await _generateRandom());
+      const id = _getEdvId(await helpers.generateRandom());
       const {config} = await storage.insertConfig(
         {actor, config: {id, ...req.body}});
       res.status(201).location(id).json(config);
@@ -153,7 +139,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       // compute invocation target
       const id = `${baseUri}/${req.originalUrl}`;
       // dynamically generate root capability for target
-      const zcap = await _generateRootCapability(id);
+      const zcap = await helpers.generateRootCapability(id);
       if(!zcap) {
         // invalid root zcap ID
         throw new BedrockError(
@@ -177,7 +163,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const edvId = _getEdvId(req.params.edvId);
       const expectedTarget = `${edvId}/documents`;
       const expectedRootCapability = `${edvId}/zcaps/documents`;
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -197,7 +183,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // check authorization
       const {docId: id} = req.params;
-      _assert128BitId(id);
+      helpers.assert128BitId(id);
       const edvId = _getEdvId(req.params.edvId);
       const docPath = `/documents/${id}`;
       const expectedTarget = [
@@ -208,7 +194,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -228,7 +214,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // check authorization
       const {docId: id} = req.params;
-      _assert128BitId(id);
+      helpers.assert128BitId(id);
       const edvId = _getEdvId(req.params.edvId);
       const docPath = `/documents/${id}`;
       const expectedTarget = [
@@ -239,7 +225,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'read'
       });
 
@@ -261,7 +247,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // check authorization
       const {docId: id} = req.params;
-      _assert128BitId(id);
+      helpers.assert128BitId(id);
       const edvId = _getEdvId(req.params.edvId);
       const docPath = `/documents/${id}`;
       const expectedTarget = [
@@ -272,7 +258,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -300,7 +286,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps/query`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'read'
       });
 
@@ -357,7 +343,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // check authorization
       const {docId} = req.params;
-      _assert128BitId(docId);
+      helpers.assert128BitId(docId);
       const edvId = _getEdvId(req.params.edvId);
       const docPath = `/documents/${docId}`;
       const expectedTarget = [
@@ -368,7 +354,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
       // TODO: add document ID to the chunk as well -- as a sanity check?
@@ -388,7 +374,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // validate `chunkIndex` and check authorization
       const {docId} = req.params;
-      _assert128BitId(docId);
+      helpers.assert128BitId(docId);
       let {chunkIndex} = req.params;
       try {
         chunkIndex = parseInt(chunkIndex, 10);
@@ -409,7 +395,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'read'
       });
 
@@ -431,7 +417,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     asyncHandler(async (req, res) => {
       // check authorization
       const {docId, chunkIndex} = req.params;
-      _assert128BitId(docId);
+      helpers.assert128BitId(docId);
       const edvId = _getEdvId(req.params.edvId);
       const docPath = `/documents/${docId}`;
       const expectedTarget = [
@@ -442,7 +428,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         `${edvId}/zcaps${docPath}`,
         `${edvId}/zcaps/documents`
       ];
-      await _authorize({
+      await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -468,7 +454,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const edvId = _getEdvId(req.params.edvId);
       const expectedTarget = `${edvId}/authorizations`;
       const expectedRootCapability = `${edvId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -488,7 +474,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const edvId = _getEdvId(req.params.edvId);
       const expectedTarget = `${edvId}/authorizations`;
       const expectedRootCapability = `${edvId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'read'
       });
 
@@ -517,7 +503,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const edvId = _getEdvId(req.params.edvId);
       const expectedTarget = `${edvId}/authorizations`;
       const expectedRootCapability = `${edvId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability, expectedAction: 'write'
       });
 
@@ -547,328 +533,3 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       }
     }));
 });
-
-// TODO: some of the following code is a target for reusability in other
-// modules and should be factored out for reuse
-
-// Note: for dereferencing `did:key` URLs
-const DOCUMENT_LOADER = async url => {
-  // TODO: move to did-key lib
-  if(url.startsWith('did:key:')) {
-    const did = url.split('#')[0];
-    const fingerprint = did.substr('did:key:'.length);
-    const publicKeyBase58 = _parsePublicKeyBase58(did);
-    const keyId = `${did}#${fingerprint}`;
-    const didDoc = {
-      '@context': SECURITY_CONTEXT_V2_URL,
-      id: did,
-      publicKey: [{
-        id: keyId,
-        // TODO: determine from parsing multibase key
-        type: 'Ed25519VerificationKey2018',
-        controller: did,
-        publicKeyBase58
-      }],
-      authentication: [keyId],
-      assertionMethod: [keyId],
-      capabilityDelegation: [keyId],
-      capabilityInvocation: [keyId]
-    };
-    const document = !url.includes('#') ? didDoc : {
-      '@context': SECURITY_CONTEXT_V2_URL,
-      ...didDoc.publicKey[0]
-    };
-    return {
-      contextUrl: null,
-      documentUrl: url,
-      document
-    };
-  }
-  const error = new Error(`Dereferencing url "${url}" is prohibited.`);
-  error.name = 'NotAllowedError';
-  error.httpStatusCode = 400;
-  throw error;
-};
-
-async function _verifyDelegation({capability, edvId}) {
-  const invocationTarget = typeof capability.invocationTarget === 'string' ?
-    capability.invocationTarget : capability.invocationTarget.id;
-  const documentLoader = extendContextLoader(async url => {
-    if(url.startsWith('did:key:')) {
-      return DOCUMENT_LOADER(url);
-    }
-
-    // dynamically generate zcap for root capability if applicable
-    const zcap = await _generateRootCapability(url);
-    if(zcap) {
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: zcap
-      };
-    }
-
-    // see if zcap is in storage
-    try {
-      const {authorization} = await brZCapStorage.authorizations.get(
-        {id: url, invocationTarget});
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: authorization.capability
-      };
-    } catch(e) {
-      if(e.name !== 'NotFoundError') {
-        throw e;
-      }
-    }
-
-    return DOCUMENT_LOADER(url);
-  });
-  const expectedRootCapability = _getExpectedRootCapability(
-    {invocationTarget, edvId});
-  const {verified, error} = await jsigs.verify(capability, {
-    suite: new Ed25519Signature2018(),
-    purpose: new CapabilityDelegation({
-      expectedRootCapability,
-      suite: new Ed25519Signature2018()
-    }),
-    documentLoader,
-    compactProof: false
-  });
-  if(!verified) {
-    throw error;
-  }
-}
-
-async function _authorize({
-  req, expectedTarget, expectedRootCapability, expectedAction
-}) {
-  // wrap document loader to always generate root zcap from config
-  // description in storage
-  const wrappedDocumentLoader = async url => {
-    if(url.startsWith('did:key:')) {
-      return DOCUMENT_LOADER(url);
-    }
-
-    // dynamically generate zcap for root capability if applicable
-    const zcap = await _generateRootCapability(url);
-    if(zcap) {
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: zcap
-      };
-    }
-
-    // see if zcap is in storage
-    try {
-      const {authorization} = await brZCapStorage.authorizations.get(
-        {id: url, invocationTarget: expectedTarget});
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: authorization.capability
-      };
-    } catch(e) {
-      if(e.name !== 'NotFoundError') {
-        throw e;
-      }
-    }
-
-    return DOCUMENT_LOADER(url);
-  };
-
-  const url = `${config.server.baseUri}${req.originalUrl}`;
-  const {method, headers} = req;
-  const result = await verifyCapabilityInvocation({
-    url, method, headers,
-    getInvokedCapability,
-    documentLoader: wrappedDocumentLoader,
-    expectedHost: config.server.host,
-    expectedTarget, expectedRootCapability, expectedAction,
-    // TODO: support RsaSignature2018 and other suites?
-    suite: [new Ed25519Signature2018()]
-  });
-  if(!result.verified) {
-    throw new BedrockError(
-      'Permission denied.', 'NotAllowedError', {
-        httpStatusCode: 400,
-        public: true
-      }, result.error);
-  }
-  return {
-    valid: result.verified,
-    ...result
-  };
-}
-
-async function getInvokedCapability({id, expectedTarget}) {
-  // if the capability is a root zcap generated by this server then its
-  // `id` will map to an invocation target; if so, dynamically generate the
-  // zcap as it is the root authority which is automatically authorized
-  const zcap = await _generateRootCapability(id);
-  if(zcap) {
-    return zcap;
-  }
-
-  // must get capability from authorizations storage
-  try {
-    const {authorization} = await brZCapStorage.authorizations.get({
-      id,
-      invocationTarget: expectedTarget
-    });
-    return authorization.capability;
-  } catch(e) {
-    if(e.name === 'NotFoundError') {
-      throw new BedrockError(
-        'Permission denied.', 'NotAllowedError', {
-          httpStatusCode: 400,
-          public: true
-        }, e);
-    }
-    throw e;
-  }
-}
-
-function _parsePublicKeyBase58(didKeyUrl) {
-  const fingerprint = didKeyUrl.substr('did:key:'.length);
-  // skip leading `z` that indicates base58 encoding
-  const buffer = base58.decode(fingerprint.substr(1));
-  // assume buffer is: 0xed 0x01 <public key bytes>
-  return base58.encode(buffer.slice(2));
-}
-
-async function _generateRootCapability(url) {
-  const result = _getInvocationTarget(url);
-  if(!result) {
-    return null;
-  }
-  const {target, edvId} = result;
-
-  // dynamically generate zcap for root capability
-  const {config} = await storage.getConfig({actor: null, id: edvId});
-  return {
-    '@context': SECURITY_CONTEXT_V2_URL,
-    id: url,
-    invocationTarget: target,
-    controller: config.controller,
-    invoker: config.invoker,
-    delegator: config.delegator
-  };
-}
-
-/**
-* Gets an `expectedRootCapability` from an `invocationTarget`. This function
-* handles special cases where the `invocationTarget` does not match the
-* `expectedRootCapability` because the `invocationTarget` cannot be expressed
-* as a zcap.
-*
-* @param {object} options - Options to use.
-* @param {string} options.invocationTarget - An invocationTarget.
-* @param {string} options.edvId - The edvId (URL to the EDV).
-*
-* @return {string} The `expectedRootCapability` for the `invocationTarget`.
- */
-function _getExpectedRootCapability({invocationTarget, edvId}) {
-  // `authorizations` endpoint cannot be expressed as a zcap, map to zcap space
-  if(invocationTarget === `${edvId}/authorizations`) {
-    return `${edvId}/zcaps/authorizations`;
-  }
-  // `documents` endpoint cannot be expressed as a zcap, map to zcap space
-  if(invocationTarget === `${edvId}/documents`) {
-    return `${edvId}/zcaps/documents`;
-  }
-  // a specific document cannot be expressed as a zcap, map to zcap space
-  // and allow root zcap to be entire collection or just the document
-  if(invocationTarget.startsWith(`${edvId}/documents/`)) {
-    const docPath = invocationTarget.substr(`${edvId}/documents/`.length);
-    return [
-      `${edvId}/zcaps/documents/${docPath}`,
-      `${edvId}/zcaps/documents`
-    ];
-  }
-  // query endpoint can't be expressed as zcap, map to zcap space and allow
-  // documents endpoint as a root zcap for queries as well
-  if(invocationTarget === `${edvId}/query`) {
-    return [
-      `${edvId}/zcaps/query`,
-      `${edvId}/zcaps/documents`
-    ];
-  }
-  // otherwise use default behavior of `invocationTarget` matches root zcap
-  return invocationTarget;
-}
-
-function _getInvocationTarget(url) {
-  // look for `/edvs/<edvId>/zcaps/`
-  const {baseUri} = config.server;
-  const baseStorageUrl = `${baseUri}${routes.edvs}/`;
-  let idx = url.indexOf(baseStorageUrl);
-  if(idx !== 0) {
-    return null;
-  }
-
-  // skip EDV ID
-  const edvIdIdx = baseStorageUrl.length;
-  idx = url.indexOf('/', edvIdIdx);
-  if(idx === -1) {
-    return null;
-  }
-  const edvId = `${baseStorageUrl}${url.substring(edvIdIdx, idx)}`;
-
-  // skip `zcaps`
-  idx = url.indexOf('zcaps/', idx + 1);
-  if(idx === -1) {
-    return null;
-  }
-
-  // valid root zcap invocation targets:
-  // `/edvs/<edvId>/documents`
-  // `/edvs/<edvId>/query`
-  // `/edvs/<edvId>/authorizations`
-  // root `/edvs/<edvId>/documents/...`
-  const path = url.substr(idx + 6 /* 'zcaps/'.length */);
-  if(!(['documents', 'query', 'authorizations'].includes(path) ||
-    (path.startsWith('documents/') && path.length > 10))) {
-    return null;
-  }
-
-  // return invocation target for the given root zcap URL
-  return {
-    target: `${edvId}/${path}`,
-    edvId
-  };
-}
-
-async function _generateRandom() {
-  // 128-bit random number, multibase encoded
-  // 0x00 = identity tag, 0x10 = length (16 bytes)
-  const buf = Buffer.concat([
-    Buffer.from([0x00, 0x10]),
-    await getRandomBytes(16)
-  ]);
-  // multibase encoding for base58 starts with 'z'
-  return 'z' + base58.encode(buf);
-}
-
-// TODO: move into JSON schema validator
-function _assert128BitId(id) {
-  try {
-    // verify ID is multibase base58-encoded 16 bytes
-    const buf = base58.decode(id.substr(1));
-    // multibase base58 (starts with 'z')
-    // 128-bit random number, multibase encoded
-    // 0x00 = identity tag, 0x10 = length (16 bytes) + 16 random bytes
-    if(!(id.startsWith('z') &&
-      buf.length === 18 && buf[0] === 0x00 && buf[1] === 0x10)) {
-      throw new Error('Invalid identifier.');
-    }
-  } catch(e) {
-    throw new BedrockError(
-      `Identifier "${id}" must be multibase, base58-encoded ` +
-      'array of 16 random bytes.',
-      'SyntaxError',
-      {public: true, httpStatusCode: 400});
-  }
-}
