+++ lib/index.js
@@ -3,7 +3,12 @@
  */
 'use strict';
 
+// load config defaults
+require('./config');
+
 require('./http');
+require('./http-revocations');
 
 // module API
 module.exports = require('./storage');
+
