+++ lib/storage.js
@@ -11,10 +11,7 @@ const database = require('bedrock-mongodb');
 const brPermission = require('bedrock-permission');
 const {promisify} = require('util');
 const brPermissionCheck = promisify(brPermission.checkPermission);
-const {BedrockError} = bedrock.util;
-
-// load config defaults
-require('./config');
+const {util: {BedrockError}} = bedrock;
 
 // module API
 const api = {};
